containers/nexus-controller/requirements.txt
fastapi
uvicorn
pydantic-ai[vertexai,openai]>=0.0.18
python-dotenv
loguru
httpx
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
openinference-instrumentation-pydantic-ai==0.1.12
openinference-instrumentation-openai==0.1.44
opentelemetry-instrumentation-httpx
prometheus_fastapi_instrumentator
slowapi
mcp
langfuse>=2.0.0
